% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{compute_county_nc}
\alias{compute_county_nc}
\title{Title}
\usage{
compute_county_nc(data, scales, county = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#nc_flat |> rename(fips = FIPS) |> compute_county_nc() |> head()
#nc_flat |> rename(fips = FIPS) |> compute_county_nc(county = "Ashe")
}
