% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{compute_county_nc_stamp}
\alias{compute_county_nc_stamp}
\title{Title}
\usage{
compute_county_nc_stamp(data, scales, county = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#nc_flat |> rename(fips = FIPS) |> compute_county_nc() |> head() |> str()
#nc_flat |> rename(fips = FIPS) |> compute_county_nc(county = "Ashe")
}
