% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{compute_county_northcarolina_stamp}
\alias{compute_county_northcarolina_stamp}
\title{Title}
\usage{
compute_county_northcarolina_stamp(data, scales, keep_county = NULL)
}
\arguments{
\item{county}{}
}
\description{
Title
}
\examples{
library(dplyr)
#northcarolina_flat |> rename(fips = FIPS) |> compute_county_northcarolina() |> head() |> str()
#northcarolina_flat |> rename(fips = FIPS) |> compute_county_northcarolina(keep_county = "Ashe")
}
