% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{compute_panel_county_centers}
\alias{compute_panel_county_centers}
\title{Title}
\usage{
compute_panel_county_centers(data, scales, keep_county = NULL)
}
\arguments{
\item{keep_county}{}
}
\description{
Title
}
\examples{
northcarolina_flat |>
  dplyr::rename(fips = FIPS) |>
  dplyr::rename(label = NAME) |>
  compute_panel_county_centers()
}
