% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{compute_panel_region_centers}
\alias{compute_panel_region_centers}
\title{Title}
\usage{
compute_panel_region_centers(data, scales, keep_region = NULL)
}
\arguments{
\item{keep_region}{}
}
\description{
Title
}
\examples{
#chile_region_flat |>
#  dplyr::rename(fips = FIPS) |>
#  dplyr::rename(label = NAME) |>
#  compute_panel_region_centers()
}
