% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{compute_region_chile}
\alias{compute_region_chile}
\title{Title}
\usage{
compute_region_chile(data, scales, keep_region = NULL)
}
\arguments{
\item{keep_region}{}
}
\description{
Title
}
\examples{
library(dplyr)
#chile_region_flat |> rename(fips = FIPS) |> compute_region_chile() |> head()
#chile_region_flat |> rename(fips = FIPS) |> compute_region_chile(keep_region = "Ashe")
}
