% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{compute_region_chile_stamp}
\alias{compute_region_chile_stamp}
\title{Title}
\usage{
compute_region_chile_stamp(data, scales, keep_region = NULL)
}
\arguments{
\item{region}{}
}
\description{
Title
}
\examples{
library(dplyr)
#chile_region_flat |> rename(fips = FIPS) |> compute_region_chile() |> head() |> str()
#chile_region_flat |> rename(fips = FIPS) |> compute_region_chile(keep_region = "Ashe")
}
