% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{geom_label_chile_region}
\alias{geom_label_chile_region}
\title{Title}
\usage{
geom_label_chile_region(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
chile_region_flat \%>\%
 ggplot() +
 aes(region_codigo = region_codigo, label = region_codigo) +
 geom_label_chile_region()

chile_region_flat \%>\%
 ggplot() +
 aes(region_codigo = region_codigo, label = region_codigo) +
 geom_sf_regionchile() +
 geom_label_chile_region()

 chile_region_flat \%>\%
 ggplot() +
 aes(region_codigo = region_codigo, label = region_codigo) +
 geom_sf_regionchile(fill = "navyblue") +
 geom_label_chile_region(color = "oldlace")

}
