% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{geom_label_northcarolina_county}
\alias{geom_label_northcarolina_county}
\title{Title}
\usage{
geom_label_northcarolina_county(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
northcarolina_flat \%>\%
 ggplot() +
 aes(fips = FIPS, label = NAME) +
 geom_label_northcarolina_county()

northcarolina_flat \%>\%
 ggplot() +
 aes(fips = FIPS, label = NAME) +
 geom_sf_countynorthcarolina() +
 geom_label_northcarolina_county()

 northcarolina_flat \%>\%
 ggplot() +
 aes(fips = FIPS, label = SID74, fill = SID74) +
 geom_sf_countynorthcarolina() +
 geom_label_northcarolina_county(color = "oldlace")

 northcarolina_flat \%>\%
 ggplot() +
 aes(fips = FIPS, fill = SID74,
     label = paste0(NAME, "\n", SID74)) +
 geom_sf_countynorthcarolina() +
 geom_label_northcarolina_county(lineheight = .7,
 size = 2, check_overlap= TRUE,
 color = "oldlace")
}
