% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_sf_countynorthcarolina}
\alias{geom_sf_countynorthcarolina}
\title{Title}
\usage{
geom_sf_countynorthcarolina(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = "NAD27",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
northcarolina_flat \%>\%
ggplot() +
aes(fips = FIPS) +
geom_sf_countynorthcarolina()
}
