% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_sf_countync}
\alias{stamp_sf_countync}
\title{Title}
\usage{
stamp_sf_countync(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = "NAD27",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot() +
stamp_sf_countync()
}
