% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_sf_regionchile}
\alias{stamp_sf_regionchile}
\title{Title}
\usage{
stamp_sf_regionchile(
  mapping = NULL,
  data = reference_full,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = "NAD27",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot() +
stamp_sf_regionchile()
}
